# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OxyGen CMS API
  version: 0.8.0
  summary: OxyGen Central Management Server API
  description: |
    API contract for OxyGen Central Management Server Phase 1.

    Current CMS schema version: 0.19.

    Phase 1 APIs cover first-run setup, local authentication/RBAC, tenant/user/group/role administration,
    manual OxyGen instance enrollment, dashboard background poller status/controls, application logs, grid preferences, application settings,
    Settings system visibility, and Phase 1.5 queue status. Protected routes use bearer authentication plus explicit permissions and Tenant/global/instance scope checks.
servers:
  - url: http://localhost:3000
    description: Local API development server
  - url: https://{host}
    description: Deployed CMS server
    variables:
      host:
        default: cms.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/app-settings/labels:
    get:
      tags: [Application Settings]
      summary: Get configurable application display labels
      responses:
        '200':
          description: Label settings.
          content:
            application/json:
              schema:
                type: object
                required: [labels]
                properties:
                  labels: { $ref: '#/components/schemas/AppLabels' }
        '401': { $ref: '#/components/responses/Error' }
    put:
      tags: [Application Settings]
      summary: Save configurable application display labels
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppLabels' }
      responses:
        '200':
          description: Saved label settings.
          content:
            application/json:
              schema:
                type: object
                required: [labels]
                properties:
                  labels: { $ref: '#/components/schemas/AppLabels' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /api/app-settings/log-retention:
    get:
      tags: [Application Settings]
      summary: Get activity retention setting
      responses:
        '200':
          description: Activity retention setting.
          content:
            application/json:
              schema:
                type: object
                required: [retention]
                properties:
                  retention: { $ref: '#/components/schemas/LogRetentionSettings' }
        '401': { $ref: '#/components/responses/Error' }
    put:
      tags: [Application Settings]
      summary: Save activity retention setting
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogRetentionSettings' }
      responses:
        '200':
          description: Saved activity retention setting.
          content:
            application/json:
              schema:
                type: object
                required: [retention]
                properties:
                  retention: { $ref: '#/components/schemas/LogRetentionSettings' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /api/app-settings/ssl-certificate-warning:
    get:
      tags: [Application Settings]
      summary: Get SSL certificate warning threshold
      description: Returns the global threshold used to label otherwise valid HTTPS certificates as Expiring Soon before they expire. Expired certificates always take precedence and display as Expired.
      responses:
        '200':
          description: SSL certificate warning setting.
          content:
            application/json:
              schema:
                type: object
                required: [sslCertificateWarning]
                properties:
                  sslCertificateWarning: { $ref: '#/components/schemas/SslCertificateWarningSettings' }
        '401': { $ref: '#/components/responses/Error' }
    put:
      tags: [Application Settings]
      summary: Save SSL certificate warning threshold
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SslCertificateWarningSettings' }
      responses:
        '200':
          description: Saved SSL certificate warning setting.
          content:
            application/json:
              schema:
                type: object
                required: [sslCertificateWarning]
                properties:
                  sslCertificateWarning: { $ref: '#/components/schemas/SslCertificateWarningSettings' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /api/app-settings/license-expiration-warning:
    get:
      tags: [Application Settings]
      summary: Get license expiration warning threshold
      description: Returns the global threshold used to label otherwise valid OxyGen licenses as Expiring Soon before they expire. Expired licenses always take precedence and display as Expired.
      responses:
        '200':
          description: License expiration warning setting.
          content:
            application/json:
              schema:
                type: object
                required: [licenseExpirationWarning]
                properties:
                  licenseExpirationWarning: { $ref: '#/components/schemas/LicenseExpirationWarningSettings' }
        '401': { $ref: '#/components/responses/Error' }
    put:
      tags: [Application Settings]
      summary: Save license expiration warning threshold
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LicenseExpirationWarningSettings' }
      responses:
        '200':
          description: Saved license expiration warning setting.
          content:
            application/json:
              schema:
                type: object
                required: [licenseExpirationWarning]
                properties:
                  licenseExpirationWarning: { $ref: '#/components/schemas/LicenseExpirationWarningSettings' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /api/app-settings/queue-schedules:
    get:
      tags: [Application Settings]
      summary: Get recurring maintenance queue schedules
      description: Returns Settings → General queue maintenance schedule settings. UI displays whole-day cadence; API persists seconds with a one-day minimum.
      responses:
        '200':
          description: Queue schedule settings.
          content:
            application/json:
              schema:
                type: object
                required: [queueSchedules]
                properties:
                  queueSchedules: { $ref: '#/components/schemas/QueueScheduleSettings' }
        '401': { $ref: '#/components/responses/Error' }
    put:
      tags: [Application Settings]
      summary: Save recurring maintenance queue schedules
      description: SystemAdmin only. Saves enabled state and cadence for maintenance schedules, reconciles BullMQ schedulers at API startup or explicit route handling, and requires `everySeconds` to be at least one day.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QueueScheduleSettingsUpdate' }
      responses:
        '200':
          description: Saved queue schedule settings.
          content:
            application/json:
              schema:
                type: object
                required: [queueSchedules]
                properties:
                  queueSchedules: { $ref: '#/components/schemas/QueueScheduleSettings' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        details: {}
...
    LogRetentionSettings:
      type: object
      required: [days]
      properties:
        days: { type: integer, minimum: 1, maximum: 3650, default: 90 }
    SslCertificateWarningSettings:
      type: object
      required: [daysBeforeExpiration]
      properties:
        daysBeforeExpiration: { type: integer, minimum: 0, maximum: 3650, default: 30, description: Number of days before expiration when an otherwise valid HTTPS certificate should display as Expiring Soon. Expired certificates always display as Expired first. }
    LicenseExpirationWarningSettings:
      type: object
      required: [daysBeforeExpiration]
      properties:
        daysBeforeExpiration: { type: integer, minimum: 0, maximum: 3650, default: 30, description: Number of days before expiration when an otherwise valid OxyGen license should display as Expiring Soon. Expired licenses always display as Expired first. }
    QueueScheduleJobKey:
      type: string
      enum:
        - database-maintenance:purge-logs
        - database-maintenance:prune-check-history
        - system-maintenance:check-application-updates
        - system-maintenance:prune-queue-history
    QueueScheduleJobSettings:
      type: object
      required: [key, queue, name, label, enabled, everySeconds]
      properties:
        key: { $ref: '#/components/schemas/QueueScheduleJobKey' }
        queue: { type: string, enum: [database-maintenance, system-maintenance] }
        name: { type: string }
        label: { type: string }
        enabled: { type: boolean }
        everySeconds: { type: integer, minimum: 86400, maximum: 2592000, description: Recurring cadence in seconds. Settings UI displays whole days and enforces a minimum of one day. }
    QueueScheduleJobUpdate:
      type: object
      required: [key, enabled, everySeconds]
      properties:
        key: { $ref: '#/components/schemas/QueueScheduleJobKey' }
        enabled: { type: boolean }
        everySeconds: { type: integer, minimum: 86400, maximum: 2592000, description: Recurring cadence in seconds; one-day minimum. }
    QueueScheduleSettings:
      type: object
      required: [jobs]
      properties:
        jobs:
          type: array
          items: { $ref: '#/components/schemas/QueueScheduleJobSettings' }
    QueueScheduleSettingsUpdate:
      type: object
      required: [jobs]
      properties:
        jobs:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/QueueScheduleJobUpdate' }
...
    AppLabels:
      type: object
      required: [tenant]
      properties:
        tenant: { type: string, minLength: 1, maxLength: 64, example: Tenant }
...